refactor(core): split grep verification helpers
